Update README.md
cin Statements and Formatting

Exercise 1

Complete the program so that a sample run inputting 22 for the number of items bought and 10.98 for the price of each item will produce the results below.

Sample run of the program:

Please input the number of items bought
22
Please input the price of each item
10.98
The total bill is $241.5600

Exercise 2

Once you have the program working, change the statement:

cout << setprecision (4) << fixed;

to

cout << setprecision(4);

Rerun the program with the same data given in Exercise 1.

How does the use of the fixed manipulator change how the setprecision function formats the decimal number? Make a note in your notes about how fixed and setprecision work.

Exercise 3

Now configure the program so that decimal numbers are formatted for two places to the right of the decimal point when outputted. You may need to put the fixed manipulator back and adjust the setprecision function.